requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=9.5.0
datasets>=2.19.0
